+++ dist2/src/utils/multi-value.js
@@ -1,7 +1,21 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.generateMultiStyleValue = void 0;
-function splitAndMatching(v, val, text, textSplited, splitValue, matchValue, styleMatchValue, multiMode) {
+function splitBaseOnMatch(matchResult, str) {
+    let reduceDefault = {
+        result: [],
+        str: str,
+    };
+    let res = matchResult.reduce((result, current) => {
+        let index = result.str.indexOf(current);
+        result.result.push(result.str.substring(0, index));
+        result.str = result.str.substring(index + current.length);
+        return result;
+    }, reduceDefault);
+    res.result.push(res.str);
+    return res.result;
+}
+function splitAndMatching(v, val, text, textSplited, splitValue, matchValue, styleMatchValue, multiMode, useSplitBaseOnMatch) {
     if (!textSplited) {
         let matchV;
         try {
@@ -19,20 +33,30 @@ function splitAndMatching(v, val, text, textSplited, splitValue, matchValue, sty
             if (!multiMode) {
                 matchValue.push(v.toString());
                 styleMatchValue.push(val);
-                splitValue = text.split(v).reduce((res, curr, index) => {
-                    if (index >= 2) {
-                        res[1] += v + curr;
-                        return res;
-                    }
-                    else {
-                        return [...res, curr];
-                    }
-                }, []);
+                if (useSplitBaseOnMatch) {
+                    splitValue = splitBaseOnMatch(matchV, text);
+                }
+                else {
+                    splitValue = text.split(v).reduce((res, curr, index) => {
+                        if (index >= 2) {
+                            res[1] += v + curr;
+                            return res;
+                        }
+                        else {
+                            return [...res, curr];
+                        }
+                    }, []);
+                }
             }
             else {
                 matchValue.push(...matchV);
                 styleMatchValue.push(...matchV.reduce((res, curr) => [...res, val], []));
-                splitValue = text.split(v);
+                if (useSplitBaseOnMatch) {
+                    splitValue = splitBaseOnMatch(matchV, text);
+                }
+                else {
+                    splitValue = text.split(v);
+                }
             }
         }
         else {
@@ -60,21 +84,33 @@ function splitAndMatching(v, val, text, textSplited, splitValue, matchValue, sty
             }
             if (match) {
                 if (!multiMode) {
-                    const splitV = sp.split(v).reduce((res, curr, index) => {
-                        if (index >= 2) {
-                            res[1] += v + curr;
-                            return res;
-                        }
-                        else {
-                            return [...res, curr];
-                        }
-                    }, []);
+                    let splitV;
+                    if (useSplitBaseOnMatch) {
+                        splitV = splitBaseOnMatch(match, sp);
+                    }
+                    else {
+                        splitV = sp.split(v).reduce((res, curr, index) => {
+                            if (index >= 2) {
+                                res[1] += v + curr;
+                                return res;
+                            }
+                            else {
+                                return [...res, curr];
+                            }
+                        }, []);
+                    }
                     newSplit.push(...splitV);
                     newStyleValue.push(val);
                     newMatchValue.push(v.toString());
                 }
                 else {
-                    const splitV = sp.split(v);
+                    let splitV;
+                    if (useSplitBaseOnMatch) {
+                        splitV = splitBaseOnMatch(match, sp);
+                    }
+                    else {
+                        splitV = sp.split(v);
+                    }
                     newSplit.push(...splitV);
                     newMatchValue.push(...match);
                     newStyleValue.push(...match.reduce((res, curr) => [...res, val], []));
@@ -101,7 +137,7 @@ function splitAndMatching(v, val, text, textSplited, splitValue, matchValue, sty
         styleMatchValue,
     };
 }
-function generateMultiStyleValue(multiStyle, text, styles, defStyleId) {
+function generateMultiStyleValue(multiStyle, text, styles, defStyleId, useSplitBaseOnMatch) {
     if (typeof multiStyle == "object") {
         let result = "";
         let matchValue = [];
@@ -115,7 +151,7 @@ function generateMultiStyleValue(multiStyle, text, styles, defStyleId) {
                 return;
             }
             else {
-                let result = splitAndMatching(v, typeof val == "string" ? val : "", text, textSplited, splitValue, matchValue, styleMatchValue, false);
+                let result = splitAndMatching(v, typeof val == "string" ? val : "", text, textSplited, splitValue, matchValue, styleMatchValue, false, useSplitBaseOnMatch);
                 textSplited = result.textSplited;
                 splitValue = result.splitValue;
                 matchValue = result.matchValue;
@@ -130,7 +166,7 @@ function generateMultiStyleValue(multiStyle, text, styles, defStyleId) {
                     if (typeof element.reg == "string") {
                         element.reg = new RegExp(element.reg, "g");
                     }
-                    let result = splitAndMatching(element.reg, element.styleId, text, textSplited, splitValue, matchValue, styleMatchValue, true);
+                    let result = splitAndMatching(element.reg, element.styleId, text, textSplited, splitValue, matchValue, styleMatchValue, true, useSplitBaseOnMatch);
                     textSplited = result.textSplited;
                     splitValue = result.splitValue;
                     matchValue = result.matchValue;
@@ -150,22 +186,29 @@ function generateMultiStyleValue(multiStyle, text, styles, defStyleId) {
             <t xml:space="preserve" >${element}</t>
         </r>`;
             }
-            result += `
+            if (matchElement.length > 0) {
+                result += `
                             <r>
            ${styles[styleID]}
             <t xml:space="preserve" >${matchElement}</t>
         </r>
                     `;
-            // result += element + matchElement;
+            }
         }
-        // result += splitValue[length];
-        result = `<si>
+        if (splitValue[length].length > 0) {
+            result = `<si>
                     ${result}
                     <r>
         ${elementStyle}
             <t>${splitValue[length]}</t>
         </r>
                     </si>`;
+        }
+        else {
+            result = `<si>
+                    ${result}
+                    </si>`;
+        }
         return result;
     }
     else {